refactor: moved CacheType out of Block to share it with vhost_user_blk
CacheType is shared between VirtioBlock and VhostUserBlock.
