+++ src/vmm/src/devices/virtio/block_common.rs
@@ -1,0 +1,19 @@
+// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use serde::{Deserialize, Serialize};
+use versionize::{VersionMap, Versionize, VersionizeError, VersionizeResult};
+use versionize_derive::Versionize;
+
+/// Configuration options for disk caching.
+// NOTICE: Any changes to this structure require a snapshot version bump.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Deserialize, Serialize, Versionize)]
+pub enum CacheType {
+    /// Flushing mechanic not will be advertised to the guest driver
+    #[default]
+    Unsafe,
+    /// Flushing mechanic will be advertised to the guest driver and
+    /// flush requests coming from the guest will be performed using
+    /// `fsync`.
+    Writeback,
+}
