+++ src/vmm/src/resources.rs
@@ -481,6 +481,7 @@ mod tests {
     use super::*;
     use crate::cpu_config::templates::{CpuTemplateType, StaticCpuTemplate};
     use crate::devices::virtio::block::VirtioBlockError;
+    use crate::devices::virtio::block_common::CacheType;
     use crate::devices::virtio::vsock::VSOCK_DEV_ID;
     use crate::resources::VmResources;
     use crate::vmm_config::boot_source::{
