+++ src/vmm/src/vmm_config/drive.rs
@@ -10,8 +10,8 @@ use serde::{Deserialize, Serialize};
 use super::RateLimiterConfig;
 pub use crate::devices::virtio::block::device::FileEngineType;
 use crate::devices::virtio::block::device::VirtioBlockConfig;
-pub use crate::devices::virtio::block::CacheType;
 use crate::devices::virtio::block::{VirtioBlock, VirtioBlockError};
+use crate::devices::virtio::block_common::CacheType;
 use crate::devices::virtio::vhost_user_block::device::{VhostUserBlock, VhostUserBlockConfig};
 use crate::devices::virtio::vhost_user_block::VhostUserBlockError;
 use crate::VmmError;
